+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Language Server: Fixed an error that would crash the language server if a `YarnCommand`- and `YarnFunction`-tagged method was declared inside a nested C# class
 - `CompilationJob.VariableDeclarations` has been renamed to `CompilationJob.Declarations`. (`VariableDeclarations` now forwards to `Declarations`, and is marked as deprecated.)
 - Language Server: completions now contain node groups, and no longer include the names of individual nodes in a group (which cannot be jumped to directly.)
+- `LineParser` now correctly detects the first `:` in a line for the implicit character name
 
 ### Removed